+++ s2gos/__init__.py
@@ -2,7 +2,8 @@
 #  Permissions are hereby granted under the terms of the Apache 2.0 License:
 #  https://opensource.org/license/apache-2-0.
 
-# noinspection PyPep8Naming
-from .version import VERSION as __version__
+from importlib.metadata import version
+
+__version__ = version("s2gos")
 
 __all__ = ["__version__"]